Moved all experiments in a separate examples folder and updated .coveragerc accordingly
