+++ examples/experiments/plot.py
